🎨 Create sub components for navbar
