+++ components/Navbar/navItems.js
@@ -1,0 +1,20 @@
+const navItems = [
+  {
+    text: "Home",
+    href: "/",
+  },
+  {
+    text: "About",
+    href: "/about",
+  },
+  {
+    text: "Resources",
+    href: "/resources",
+  },
+  {
+    text: "Contact",
+    href: "/contact",
+  },
+];
+
+export default navItems;
